requirements.txt
path==16.3.0
mtranslate
psycopg2-binary
python-dotenv
Flask
flask-cors
flask_wtf
mtranslate